+++ CHANGES.rst
@@ -22,7 +22,12 @@ Release Date not Decided
     in the body. They were originally omitted when implementing
     ``If-Match`` (`#1233`_), but the spec doesn't seem to disallow it.
     (`#1231`_, `#1255`_)
+-   :meth:`MapAdapter.build() <routing.MapAdapter.build>` can be passed
+    a :class:`~datastructures.MultiDict` to represent multiple values
+    for a key. It already did this when passing a dict with a list
+    value. (`#724`_)
 
+.. _`#724`: https://github.com/pallets/werkzeug/pull/724
 .. _`#1231`: https://github.com/pallets/werkzeug/issues/1231
 .. _`#1233`: https://github.com/pallets/werkzeug/pull/1233
 .. _`#1252`: https://github.com/pallets/werkzeug/pull/1252